API creacion de Tarjetas

+++ APICloudCash/APICloudCash/Tarjetas.cs
@@ -23,7 +23,7 @@ namespace APICloudCash
     
         public long id_Tarjeta { get; set; }
         public long id_Cliente { get; set; }
-        public int numeroTarjeta { get; set; }
+        public long numeroTarjeta { get; set; }
         public string nombrePoseedor { get; set; }
         public System.DateTime fechaVencimiento { get; set; }
         public short cvc { get; set; }
